Fix instantiation of a fake DockerDriver
* Move mock client to nova.tests.virt.docker.mock_client
* Allow overriding of the client class in DockerDriver.__init__
* Mock out the one DockerDriver method that depended on 'fake'
* Remove DockerDriver.fake attribute
